selfimproving/requirements.txt
prompt-learning-enhanced @ git+https://github.com/arize-ai/prompt-learning.git
arize-phoenix>=4.0.0
openinference-instrumentation-google-genai>=0.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
rich>=13.0.0
